feat: implement DeviceLooperPanel for GP-200's built-in looper and refactor related components
